add error control
add error control that basically changes data that's bad into
data that's good. For example if you put that it is the 13th
month it will not use that number, it will default back to
the current month. If the user leaves the field blank, it
will default to the component of 1/1/1900 00:00:00.