rework viewmodel with private setters
rework viewmodel so it has private mutablestateof items with
functions for setters that the onClick and onValueChange
events of the composeables can use. Added validFloat func
to make sure the numeric numToAdd is numeric. What you can't
see is that in the previous versions, instead of functions
for doing stuff the viewModel was creating onClick events, it
was terrible. Anyway after the viewmodel was updated all of
the composables had to be updated to match.