update navigation
update navigation so that calling Help screen doesn't make the
back button optiosn weird. but, that made it so every time
you click back button, it goes to home screen. That's OK since
the state of the screens are saved so if you were on DateDiff
you can just click DateDiff again. This is not perfect but
it's acceptable for now.